Split Cli.scala into xxxRunner.scala files
